refactor Agent related components, to move them out of Pages/ManageUsers into Pages/ManageAgents
