+++ src/dashboard/src/Pages/ManageAgents/ManageAgents.js
@@ -3,7 +3,7 @@ import useStyle from './styles'
 import MUIDataTable from 'mui-datatables'
 import {toast} from "react-toastify";
 import Button from "@material-ui/core/Button";
-import PageTitle from "../../Components/PageTitle";
+import PageTitle from "../../Components/PageTitle/PageTitle";
 import {deleteAgent, getAgentList} from "../../Api/api_agents";
 import NewAgentModal from "./NewAgentModal";
 import AgentDetailModal from "./AgentDetailModal";
